refactor: Dot 컴포넌트 분리
